mcp-lab/requirements.txt
azure-functions
fastmcp
msal
requests
werkzeug<3.1.2
pydantic>=2.0.0
openapi-core>=0.19.0
